% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{get_utm}
\alias{get_utm}
\title{Return UTM zone (EPSG code) based on geographic coordinates}
\usage{
get_utm(x, y)
}
\arguments{
\item{x}{Numeric scalar.
Longitude coordinate.}

\item{y}{Numeric scalar.
Latitude coordinate.}
}
\description{
Return UTM zone (EPSG code) based on geographic coordinates
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
