% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{point2utm}
\alias{point2utm}
\title{Return UTM zone (EPSG code) based on a point layer}
\usage{
point2utm(point)
}
\arguments{
\item{point}{Point object of class \code{sf}.}
}
\description{
Return UTM zone (EPSG code) based on a point layer
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
