% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{wsd_unnested}
\alias{wsd_unnested}
\title{Delineate unnested watersheds}
\usage{
wsd_unnested(
  outlet,
  id_col = NULL,
  f_dir,
  str_grid = NULL,
  snap = TRUE,
  snap_dist = 5
)
}
\arguments{
\item{outlet}{Outlet point layer of class \code{sf}.}

\item{id_col}{Column name specifying outlet id.
This column information will be appended to the output polygon layer.}

\item{f_dir}{Flow direction raster of class \code{SpatRaster}.}

\item{str_grid}{Stream grid raster of class \code{SpatRaster}.
Required if \code{snap = TRUE}.}

\item{snap}{Logical. Whether snapping of outlet points should be performed.
Defaults to \code{TRUE}.}

\item{snap_dist}{Numeric. Distance threshold for snapping points to stream grid.
Measured in the unit of input raster files. Defaults to \code{5}.}
}
\value{
A polygon layer of class \code{sf} with one row per outlet point.
If \code{id_col} is supplied, the output includes the identifier column and
coordinates of the original (\code{x0}, \code{y0}) and snapped
(\code{x}, \code{y}) outlet points.
}
\description{
Delineate unnested watersheds for multiple outlet points using a
D8 flow direction raster.
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
