% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{arc2d8}
\alias{arc2d8}
\title{Convert ArcGIS flow direction to D8 flow direction}
\usage{
arc2d8(x)
}
\arguments{
\item{x}{Flow direction raster of class \code{SpatRaster}}
}
\description{
Convert ArcGIS flow direction to D8 flow direction
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
