% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{flow2grid}
\alias{flow2grid}
\title{Convert flow accumulation raster to stream grid}
\usage{
flow2grid(f_acc, threshold, output)
}
\arguments{
\item{f_acc}{Flow accumulation raster of class \code{SpatRaster}.}

\item{threshold}{Numeric.
Threshold value for minimum stream grid.
The unit inherits from input the flow accumulation layer.}

\item{output}{Character.
Output file path.}
}
\description{
Convert flow accumulation raster to stream grid
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
