% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{get_key}
\alias{get_key}
\title{Return file name keys for MERITHYDRO rasters}
\usage{
get_key(shape)
}
\arguments{
\item{shape}{\code{sf} polygon object.
This object defines the extent.}
}
\description{
Return file name keys for MERITHYDRO rasters
}
\section{MERITHYDRO}{
 https://hydro.iis.u-tokyo.ac.jp/~yamadai/MERIT_Hydro/
}

\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
