% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_tf}
\alias{get_tf}
\title{Utility function: Return the terminal cell of a given polygon}
\usage{
get_tf(data, shape, mode)
}
\arguments{
\item{data}{Dataframe.
This object must list the maximum and minimum of coordinates
(latitude or longitude)}

\item{shape}{Polygon object of class \code{sf}.
This object defines the extent.}

\item{mode}{Character.
This argument specifies which terminal coordinate of the shape polygon
should be used. Either \code{xmin, xmax, ymin, ymax}}
}
\description{
Utility function: Return the terminal cell of a given polygon
}
