% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{grid2stream}
\alias{grid2stream}
\title{Convert stream grid to vector stream}
\usage{
grid2stream(f_dir, str_grid, output, set_crs = TRUE)
}
\arguments{
\item{f_dir}{Flow direction raster of class \code{SpatRaster}.}

\item{str_grid}{Stream grid raster of class \code{SpatRaster}.}

\item{output}{Character.
File path for output stream vector.}

\item{set_crs}{Logical.
Whether output file should inherit CRS from the input flow direction raster.}
}
\description{
Convert stream grid to vector stream
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
