% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{rast2poly}
\alias{rast2poly}
\title{Convert raster to polygon(s)}
\usage{
rast2poly(x, simplify = TRUE, keep = 0.05)
}
\arguments{
\item{x}{Raster of class \code{SpatRaster}}

\item{simplify}{Logical.
Whether output polygons are simplified or not.}

\item{keep}{Numeric.
Proportion of vertices kept after polygon simplifications.
Ignored if \code{simplify = FALSE}}
}
\description{
Convert raster to polygon(s)
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
