% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{wsd_nested}
\alias{wsd_nested}
\title{Delineate nested watersheds}
\usage{
wsd_nested(
  outlet,
  id_col = NULL,
  f_dir,
  str_grid = NULL,
  snap = TRUE,
  snap_dist = 5
)
}
\arguments{
\item{outlet}{Outlet point layer of class \code{sf}.}

\item{id_col}{Column name specifying outlet id.
This column information will be appended to the output polygon layer.}

\item{f_dir}{Flow direction raster of class \code{SpatRaster}.}

\item{str_grid}{Stream grid raster of class \code{SpatRaster}.
Required if \code{snap = TRUE}.}

\item{snap}{Logical. Whether snapping of outlet points should be performed.
Defaults to \code{TRUE}.}

\item{snap_dist}{Numeric. Distance threshold for snapping points to stream grid.
Measured in the unit of input raster files. Defaults to \code{5}.}
}
\value{
A named list with:
\describe{
\item{watershed}{Polygon layer of class \code{sf}, one row per outlet.}
\item{outlet}{Snapped outlet points of class \code{sf}.}
}
}
\description{
Delineate nested watersheds for multiple outlet points using a D8 flow
direction raster.
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
