% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapbox.R
\name{wsd_unnested}
\alias{wsd_unnested}
\title{Delineate unnested watersheds}
\usage{
wsd_unnested(
  outlet,
  id_col,
  f_dir,
  str_grid = NULL,
  snap = TRUE,
  snap_dist = 5,
  export = FALSE,
  output_dir = "watershed",
  filename = "watershed",
  file_ext = "gpkg",
  keep_outlet = FALSE
)
}
\arguments{
\item{outlet}{Outlet point layer of class \code{sf}.}

\item{id_col}{Column name specifying outlet id.
This column information will be appended to the output polygon layer.}

\item{f_dir}{Flow direction raster of class \code{SpatRaster}.}

\item{str_grid}{Stream grid raster of class \code{SpatRaster}.}

\item{snap}{Logical.
Whether snapping of outlet points should be performed.}

\item{snap_dist}{Numeric.
Distance threshold for snapping points to stream grid.
Measured in the unit of input raster files.}

\item{export}{Logical.
Whether output polygons should be exported.}

\item{output_dir}{Character.
Specify a directly name for output.
Ignored if \code{export = FALSE}.}

\item{filename}{Character.
Specify a file name of output watershed polygons.
Ignored if \code{export = FALSE}.}

\item{file_ext}{Character.
Specify a file extension of output files.
Either \code{"gpkg"} or \code{"shp"}.
Ignored if \code{export = FALSE}.}

\item{keep_outlet}{Logical.
Whether a snapped outlet layer should be exported.
Ignored if \code{export = FALSE}.}
}
\description{
Delineate unnested watersheds
}
\author{
Akira Terui, \email{hanabi0111@gmail.com}
}
